requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
flask>=3.0.0
flask-cors>=4.0.0

